+++ src/google_apps_script/functions/calculate_available_time.ts
@@ -2,7 +2,7 @@ import { AvailableTimeCalculateInteractor } from '../../domain/applications/avai
 import { AvailableTimeRepository } from '../../infrastructure/available_times/available_time_repository';
 import { IsoWeekRepository } from '../../infrastructure/iso_weeks/iso_week_repository';
 import { UserSettingRepository } from '../../infrastructure/user_settings/user_setting_repository';
-import { AvailableTimeCalculateController } from '../../webhook_app/available_time/controllers/calculate/available_time_calculate_controller';
+import { AvailableTimeCalculateController } from '../../webhook_app/controllers/available_time/calculate/available_time_calculate_controller';
 
 function calculateAvailableTime(): void {
   const availableTimeRepository = new AvailableTimeRepository();
