+++ src/google_apps_script/functions/notify_available_time.ts
@@ -1,7 +1,7 @@
 import { AvailableTimeNotifyInteractor } from '../../domain/applications/available_time/available_time_notify_interactor';
 import { UserSettingRepository } from '../../infrastructure/user_settings/user_setting_repository';
-import { AvailableTimeNotifyController } from '../../webhook_app/available_time/controllers/notify/available_time_notify_controller';
-import { NotifyPresenter } from '../../webhook_app/common/presenters/notify/notify_presenter';
+import { AvailableTimeNotifyController } from '../../webhook_app/controllers/available_time/notify/available_time_notify_controller';
+import { NotifyPresenter } from '../../webhook_app/presenters/notify/notify_presenter';
 
 function notifyAvailableTime(): void {
   const userSettingRepository = new UserSettingRepository();
