--- src/webhook_app/common/view_models/notify/notify_view_model.ts
@@ -1,8 +1,0 @@
-import { AvailableTimeSlackNotifyView } from '../../../available_time/views/slack/notify/available_time_slack_notify_view';
-
-export class NotifyViewModel {
-  notify(message: string): void {
-    const slack = new AvailableTimeSlackNotifyView();
-    slack.notify(message);
-  }
-}
