--- src/webhook_app/common/view_models/reply/reply_view_model.ts
@@ -1,8 +1,0 @@
-import { UserSlackReplyView } from '../../../user/views/slack/reply/user_slack_reply_view';
-
-export class ReplyViewModel {
-  reply(message: string): string {
-    const slack = new UserSlackReplyView();
-    return slack.reply(message);
-  }
-}
