--- src/webhook_app/measurement/views/slack/reply/measurement_slack_reply_view.ts
@@ -1,1 +1,0 @@
-export class MeasurementSlackReplyView {}
