+++ src/webhook_app/presenters/notify/notify_presenter.ts
@@ -1,4 +1,4 @@
-import { NotifyPresenterInterface } from '../../../../use_case/common/notify_presenter_interface';
+import { NotifyPresenterInterface } from '../../../use_case/common/notify_presenter_interface';
 import { NotifyViewModel } from '../../view_models/notify/notify_view_model';
 
 export class NotifyPresenter implements NotifyPresenterInterface {
