+++ src/webhook_app/presenters/reply/reply_presenter.ts
@@ -1,4 +1,4 @@
-import { ReplyPresenterInterface } from '../../../../use_case/common/reply_presenter_interface';
+import { ReplyPresenterInterface } from '../../../use_case/common/reply_presenter_interface';
 import { ReplyViewModel } from '../../view_models/reply/reply_view_model';
 
 export class ReplyPresenter implements ReplyPresenterInterface {
