--- src/webhook_app/summary_report/views/slack/reply/summary_report_slack_reply_view.ts
@@ -1,1 +1,0 @@
-export class SummaryReportSlackReplyView {}
