+++ src/webhook_app/view_models/notify/notify_view_model.ts
@@ -1,0 +1,8 @@
+import { SlackNotifyView } from '../../views/slack/notify/slack_notify_view';
+
+export class NotifyViewModel {
+  notify(message: string): void {
+    const slack = new SlackNotifyView();
+    slack.notify(message);
+  }
+}
