+++ src/webhook_app/view_models/reply/reply_view_model.ts
@@ -1,0 +1,8 @@
+import { SlackReplyView } from '../../views/slack/reply/slack_reply_view';
+
+export class ReplyViewModel {
+  reply(message: string): string {
+    const slack = new SlackReplyView();
+    return slack.reply(message);
+  }
+}
