+++ src/webhook_app/views/slack/reply/slack_reply_view.ts
@@ -1,4 +1,4 @@
-export class UserSlackReplyView {
+export class SlackReplyView {
   reply(message: string): string {
     return message;
   }
